requirements.txt
# OCR Dependencies
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# LLM Integration
ollama>=0.1.0

# Markdown Processing
markdown>=3.5
python-markdown-math>=0.8

# Utilities
click>=8.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
